+++ src/tmpl/router.js
@@ -12,14 +12,14 @@ var Router_LLoc = {
 };
 var Router_TrimHashReg = /(?:^.*\/\/[^\/]+|#.*$)/gi;
 var Router_TrimQueryReg = /^[^#]*#?!?/;
-var GetParam = function (key, params) {
+var GetParam = function(key, params) {
     params = this[G_PARAMS];
     return params[key] || G_EMPTY;
 };
 var Router_Edge;
 /*#if(!modules.forceEdgeRouter){#*/
 var Router_Hashbang = G_HashKey + '!';
-var Router_UpdateHash = function (path, replace) {
+var Router_UpdateHash = function(path, replace) {
     path = Router_Hashbang + path;
     if (replace) {
         Router_WinLoc.replace(path);
@@ -27,7 +27,7 @@ var Router_UpdateHash = function (path, replace) {
         Router_WinLoc.hash = path;
     }
 };
-var Router_Update = function (path, params, loc, replace, silent, lQuery) {
+var Router_Update = function(path, params, loc, replace, silent, lQuery) {
     path = G_ToUri(path, params, lQuery);
     if (path != loc.srcHash) {
         Router_Silent = silent;
@@ -35,10 +35,10 @@ var Router_Update = function (path, params, loc, replace, silent, lQuery) {
     }
 };
 /*#if(modules.tipRouter){#*/
-var Router_Bind = function () {
+var Router_Bind = function() {
     var lastHash = Router_Parse().srcHash;
     var newHash, suspend;
-    G_DOMEventLibBind(G_WINDOW, 'hashchange', function (e, loc, resolve) {
+    G_DOMEventLibBind(G_WINDOW, 'hashchange', function(e, loc, resolve) {
         if (suspend) {
             /*#if(modules.tipLockUrlRouter){#*/
             Router_UpdateHash(lastHash);
@@ -48,7 +48,7 @@ var Router_Bind = function () {
         loc = Router_Parse();
         newHash = loc.srcHash;
         if (newHash != lastHash) {
-            resolve = function () {
+            resolve = function() {
                 e.p = 1;
                 lastHash = newHash;
                 suspend = G_EMPTY;
@@ -56,7 +56,7 @@ var Router_Bind = function () {
                 Router_Diff();
             };
             e = {
-                reject: function () {
+                reject: function() {
                     e.p = 1;
                     suspend = G_EMPTY;
                     /*#if(!modules.tipLockUrlRouter){#*/
@@ -64,7 +64,7 @@ var Router_Bind = function () {
                     /*#}#*/
                 },
                 resolve: resolve,
-                prevent: function () {
+                prevent: function() {
                     suspend = 1;
                     /*#if(modules.tipLockUrlRouter){#*/
                     Router_UpdateHash(lastHash);
@@ -77,19 +77,19 @@ var Router_Bind = function () {
             }
         }
     });
-    G_WINDOW.onbeforeunload = function (e) {
+    G_WINDOW.onbeforeunload = function(e, te, msg) {
         e = e || G_WINDOW.event;
-        var te = {};
+        te = {};
         Router.fire('pageunload', te);
-        if (te.msg) {
-            if (e) e.returnValue = te.msg;
-            return te.msg;
+        if ((msg = te.msg)) {
+            if (e) e.returnValue = msg;
+            return msg;
         }
     };
     Router_Diff();
 };
 /*#}else{#*/
-var Router_Bind = function () {
+var Router_Bind = function() {
     G_DOMEventLibBind(G_WINDOW, 'hashchange', Router_Diff);
     Router_Diff();
 };
@@ -102,11 +102,11 @@ if (WinHistory.pushState) {
     /*#}#*/
     Router_Edge = 1;
     var Router_DidUpdate;
-    var Router_UpdateState = function (path, replace) {
+    var Router_UpdateState = function(path, replace) {
         WinHistory[replace ? 'replaceState' : 'pushState'](G_NULL, G_NULL, path);
     };
     var Router_Popstate;
-    var Router_Update = function (path, params, loc, replace, silent) {
+    var Router_Update = function(path, params, loc, replace, silent) {
         path = G_ToUri(path, params);
         if (path != loc.srcQuery) {
             Router_Silent = silent;
@@ -119,11 +119,11 @@ if (WinHistory.pushState) {
         }
     };
     /*#if(modules.tipRouter){#*/
-    var Router_Bind = function () {
+    var Router_Bind = function() {
         var initialURL = Router_WinLoc.href;
         var lastHref = initialURL;
         var newHref, suspend;
-        G_DOMEventLibBind(G_WINDOW, 'popstate', Router_Popstate = function (f, e, resolve) {
+        G_DOMEventLibBind(G_WINDOW, 'popstate', Router_Popstate = function(f, e, resolve) {
             newHref = Router_WinLoc.href;
             var initPop = !Router_DidUpdate && newHref == initialURL;
             Router_DidUpdate = 1;
@@ -135,14 +135,15 @@ if (WinHistory.pushState) {
                 return;
             }
             if (newHref != lastHref) {
-                resolve = function () {
+                resolve = function() {
                     e.p = 1;
                     suspend = G_EMPTY;
-                    if (!f) Router_UpdateState(lastHref = newHref);
+                    lastHref = newHref;
+                    if (!f) Router_UpdateState(newHref);
                     Router_Diff();
                 };
                 e = {
-                    reject: function () {
+                    reject: function() {
                         suspend = G_EMPTY;
                         e.p = 1;
                         /*#if(!modules.tipLockUrlRouter){#*/
@@ -150,7 +151,7 @@ if (WinHistory.pushState) {
                         /*#}#*/
                     },
                     resolve: resolve,
-                    prevent: function () {
+                    prevent: function() {
                         suspend = 1;
                         /*#if(modules.tipLockUrlRouter){#*/
                         Router_UpdateState(lastHref);
@@ -163,21 +164,21 @@ if (WinHistory.pushState) {
                 }
             }
         });
-        G_WINDOW.onbeforeunload = function (e) {
+        G_WINDOW.onbeforeunload = function(e, te, msg) {
             e = e || G_WINDOW.event;
-            var te = {};
+            te = {};
             Router.fire('pageunload', te);
-            if (te.msg) {
-                if (e) e.returnValue = te.msg;
-                return te.msg;
+            if ((msg = te.msg)) {
+                if (e) e.returnValue = msg;
+                return msg;
             }
         };
         Router_Diff();
     };
     /*#}else{#*/
-    var Router_Bind = function () {
+    var Router_Bind = function() {
         var initialURL = Router_WinLoc.href;
-        G_DOMEventLibBind(G_WINDOW, 'popstate', function () {
+        G_DOMEventLibBind(G_WINDOW, 'popstate', function() {
             var initPop = !Router_DidUpdate && Router_WinLoc.href == initialURL;
             Router_DidUpdate = 1;
             if (initPop) return;
@@ -200,7 +201,7 @@ var Router_PNR_Rewrite;
 /*#if(modules.updateTitleRouter){#*/
 var DefaultTitle = document.title;
 /*#}#*/
-var Router_AttachViewAndPath = function (loc, view) {
+var Router_AttachViewAndPath = function(loc, view) {
     if (!Router_PNR_Routers) {
         Router_PNR_Routers = Magix_Cfg.routes || {};
         Router_PNR_UnmatchView = Magix_Cfg.unmatchView;
@@ -250,7 +251,7 @@ var Router_AttachViewAndPath = function (loc, view) {
     }
 };
 
-var Router_GetChged = function (oldLocation, newLocation) {
+var Router_GetChged = function(oldLocation, newLocation) {
     var oKey = oldLocation.href;
     var nKey = newLocation.href;
     var tKey = oKey + G_SPLITER + nKey;
@@ -294,7 +295,7 @@ var Router_GetChged = function (oldLocation, newLocation) {
     }
     return result;
 };
-var Router_Parse = function (href) {
+var Router_Parse = function(href) {
     href = href || Router_WinLoc.href;
 
     var result = Router_HrefCache.get(href),
@@ -325,7 +326,7 @@ var Router_Parse = function (href) {
     }
     return result;
 };
-var Router_Diff = function () {
+var Router_Diff = function() {
     var location = Router_Parse();
     var changed = Router_GetChged(Router_LLoc, Router_LLoc = location);
     if (!Router_Silent && changed.a) {
@@ -335,7 +336,7 @@ var Router_Diff = function () {
             document.title = location.title || DefaultTitle;
         }
         /*#}#*/
-        Router.fire('changed', /*#if(modules.updateTitleRouter){#*/ Router_LastChanged /*#}else{#*/ Router_LastChanged = changed.b /*#}#*/);
+        Router.fire('changed', /*#if(modules.updateTitleRouter){#*/ Router_LastChanged /*#}else{#*/ Router_LastChanged = changed.b /*#}#*/ );
     }
     Router_Silent = 0;
     if (DEBUG) {
@@ -394,7 +395,7 @@ var Router = G_Mix({
      *
      * //凡是带path的修改地址栏，都会把原来地址栏中的参数丢弃
      */
-    to: function (pn, params, replace, silent) {
+    to: function(pn, params, replace, silent) {
         if (!params && G_IsObject(pn)) {
             params = pn;
             pn = G_EMPTY;