+++ src/webpack/magix.js
@@ -6,7 +6,7 @@ module.exports = (function() {
     var G_NOOP = function() {};
     Inc('../tmpl/variable');
     Inc('../tmpl/cache');
-    //var G_IsFunction = $.isFunction;
+    var G_IsFunction = $.isFunction;
     /*#if(modules.defaultView){#*/
     var coreDefaultView;
     /*#}#*/
@@ -53,11 +53,16 @@ module.exports = (function() {
         }
         checkCount();
     };
+    var G_Define = function(mId, value) {
+        var views = Magix_Cfg.views || G_NOOP;
+        views[mId] = value;
+    };
     Inc('../tmpl/extend');
-    var G_HTML = function(node, html) {
+    var G_HTML = function(node, html, vId) {
         $(node).html(html);
         G_DOC.triggerHandler({
             type: 'htmlchange',
+            vId: vId,
             target: node
         });
     };
@@ -68,6 +73,25 @@ module.exports = (function() {
         e.eventTarget = d.e;
         G_ToTry(d.f, e, d.v);
     };
+    /*#if(modules.eventEnterLeave){#*/
+    var Specials = {
+        mouseenter: 1,
+        mouseleave: 1,
+        pointerenter: 1,
+        pointerleave: 1
+    };
+    var G_DOMEventLibBind = function(node, type, cb, remove, scope, selector) {
+        if (scope) {
+            type += '.' + scope.i;
+        }
+        selector = Specials[type] === 1 ? '[mx-' + type + ']' : G_EMPTY;
+        if (remove) {
+            $(node).off(type, selector, cb);
+        } else {
+            $(node).on(type, selector, scope, cb);
+        }
+    };
+    /*#}else{#*/
     var G_DOMEventLibBind = function(node, type, cb, remove, scope) {
         if (scope) {
             type += '.' + scope.i;
@@ -78,6 +102,7 @@ module.exports = (function() {
             $(node).on(type, scope, cb);
         }
     };
+    /*#}#*/
 
     Inc('../tmpl/safeguard');
     Inc('../tmpl/magix');
@@ -94,13 +119,23 @@ module.exports = (function() {
     /*#}#*/
     Inc('../tmpl/vframe');
     /*#if(modules.nodeAttachVframe){#*/
-    $.fn.invokeView = function() {
-        var vf = this.prop('vframe'),
-            returned;
-        if (vf) {
-            returned = vf.invoke.apply(vf, arguments);
+    $.fn.invokeView = function(name, args) {
+        var l = this.length;
+        if (l) {
+            var e = this[0];
+            var vf = e.vframe;
+            if (args === undefined) {
+                return vf && vf.invoke(name);
+            } else {
+                for (var i = 0; i < l; i++) {
+                    e = this[i];
+                    vf = e.vframe;
+                    if (vf) {
+                        vf.invoke(name, args);
+                    }
+                }
+            }
         }
-        return returned;
     };
     /*#}#*/
 
@@ -109,11 +144,12 @@ module.exports = (function() {
     Inc('../tmpl/tmpl');
     /*#if(modules.updaterIncrement){#*/
     Inc('../tmpl/increment');
-    var Updater_Increment = function(node, html) {
+    var Updater_Increment = function(node, html, vId) {
         Increment(node, html);
         G_DOC.triggerHandler({
             type: 'htmlchange',
-            target: node
+            target: node,
+            vId: vId
         });
     };
     /*#}#*/